Give the camera open lock its own module, so importing a camera driver does not pull it in
Nothing about the lock is package-level: one driver takes it, and `__init__.py` was empty before. It sat there only because it has to import where `pyzed` does not, which a sibling module satisfies as well.

Ticket: Positronic-Robotics/internal#258 #open
